move CertValidation to cert_checker.go
